Integrating Miles and Craig Code into main scripts

+++ Lab09/EncoderDriver.py
@@ -99,7 +99,7 @@ class EncoderDriver:
                 self.position = (self.position + (self.delta + self.overflow))
             elif(self.delta > 0):
                 # In the case of underflow
-                self.position = (self.position + (self.delta - self.overflow))
+                self.position = int(self.position + (self.delta - self.overflow))
         
         # Current count becomes previous count for next iteration
         self.prev_count = self.curr_count
@@ -145,7 +145,7 @@ class EncoderDriver:
         '''
     
         ## Position of encoder in degrees
-        theta = ticks*(1/self.PPC)*(1/self.CPR)*(360/1)
+        theta = int(ticks*(1/self.PPC)*(1/self.CPR)*(360/1))
     
         return theta
